Added AI training plans into different Agent Scripts.
