+++ Assets/Prefabs/Main Camera 1.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3ca8aabc79dd6aa40a16a79befaf6f0c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
