--- Assets/Prefabs/Main Camera.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: dfb0c2e612d86544dbf8c530ae29bab3
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
