+++ Assets/Prefabs/StarField.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e2405e296e0a88940a1088dd9c42b8ec
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
