+++ Assets/Scenes/TrainingScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9bbb74d245736f04c92442f6fff6dde1
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
