+++ Assets/Scripts/GameManager.cs.meta
@@ -3,7 +3,18 @@ guid: e2b20f6dac4a0eb4ea9e8b382c79efdd
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
-  defaultReferences: []
+  defaultReferences:
+  - ballPrefab: {fileID: 3216065446569584671, guid: 26b0ac11519856847bfd17a75ad2a288, type: 3}
+  - playerPrefab: {fileID: 5106813669491007435, guid: 938da22dafeee0c459ed2bbefaa5b523, type: 3}
+  - scoreText: {instanceID: 0}
+  - ballsText: {instanceID: 0}
+  - levelText: {instanceID: 0}
+  - TotalScoreText: {instanceID: 0}
+  - panelMenu: {instanceID: 0}
+  - panelPlay: {instanceID: 0}
+  - panelLevelCompleted: {instanceID: 0}
+  - panelGameOver: {instanceID: 0}
+  - currentBall: {instanceID: 0}
   executionOrder: 0
   icon: {instanceID: 0}
   userData:
