--- Assets/Scripts/PaddleAgent.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: de627a477339db847a2cb6873e09b7b5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
